moved components, created header title component
