+++ src/components/homepage/header/search_group/SearchHeader.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function SearchHeader() {
+  return (
+    <h1>FIND YOUR MOVIE</h1>
+  )
+}
